[ft #[163518687]] remove social login
[ft #[163518687]] fixing hound CI errors

[ft #[163518687]] fixing hound CI errors

[ft #[163518687] refractor code and tests for social authentication

[ft  #[163518687] delete unused files for social authentication
